requirements.txt
asgiref==3.8.1
attrs==24.3.0
black==24.10.0
click==8.1.8
colorama==0.4.6
Django==5.1.4
django-debug-toolbar==5.0.0a0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
Faker==33.1.0
flake8==7.1.1
flake8-quotes==3.4.0
flake8-variables-names==0.0.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pep8-naming==0.13.2
pillow==11.0.0
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
